Upgrade to latest primogenitor
This upgrades the parent pom to the latest version. This introduces
a new requirement on JDK 11, and the project now produces
JDK 11 bytecode. It switches to the bnd-maven-plugin from the
maven-bundle-plugin as this should give somewhat better cross-IDE
support with regards to OSGi development. It also uses the new OSGi
bundle annotations to export packages rather than specifying the
configuration in pom files. Unfortunately, the maven-dependency-plugin
is temporarily disabled due to an incompatibility with JDK 11 bytecode.